resolve

+++ src/lib/gundb/UserStorageClass.js
@@ -4,7 +4,6 @@ import Mutex from 'await-mutex'
 import SEA from 'gun/sea'
 import find from 'lodash/find'
 import flatten from 'lodash/flatten'
-import get from 'lodash/get'
 import isEqual from 'lodash/isEqual'
 import keys from 'lodash/keys'
 import maxBy from 'lodash/maxBy'
@@ -14,12 +13,13 @@ import takeWhile from 'lodash/takeWhile'
 import toPairs from 'lodash/toPairs'
 import values from 'lodash/values'
 import isEmail from 'validator/lib/isEmail'
+import moment from 'moment'
 import Config from '../../config/config'
 import API from '../API/api'
-
 import pino from '../logger/pino-logger'
 import isMobilePhone from '../validators/isMobilePhone'
 import defaultGun from './gundb'
+import UserProperties from './UserPropertiesClass'
 import { getUserModel, type UserModel } from './UserModel'
 
 const logger = pino.child({ from: 'UserStorage' })
@@ -91,6 +91,15 @@ export type FeedEvent = {
 }
 
 /**
+ * Survey details
+ */
+export type SurveyDetails = {
+  amount: string,
+  reason: string,
+  survey: string,
+}
+
+/**
  * Blockchain transaction event data
  */
 export type TransactionEvent = FeedEvent & {
@@ -106,20 +115,95 @@ export type TransactionEvent = FeedEvent & {
 }
 
 export const welcomeMessage = {
-  id: '0',
+  id: '1',
+  type: 'welcome',
+  date: new Date().toString(),
+  status: 'completed',
+  data: {
+    customName: 'Welcome to GoodDollar',
+    subtitle: 'Welcome to GoodDollar',
+    receiptData: {
+      from: '0x0000000000000000000000000000000000000000',
+    },
+    reason:
+      'GoodDollar is a digital coin with built-in\nbasic income. Start collecting your income by claiming GoodDollars every day.',
+    endpoint: {
+      fullName: 'Welcome to GoodDollar',
+    },
+  },
+}
+
+export const welcomeMessageOnlyEtoro = {
+  id: '1',
   type: 'welcome',
   date: new Date().toString(),
   status: 'completed',
   data: {
     customName: 'Welcome to GoodDollar!',
-    subtitle: 'Start claiming free G$',
+    subtitle: 'Welcome to GoodDollar',
     receiptData: {
       from: '0x0000000000000000000000000000000000000000',
     },
     reason:
-      'GoodDollar is a payment system with a built-in small basic income based on blockchain technology.\nLet’s change the world, for good.',
+      'Start collecting your income by claiming GoodDollars every day. Since this is a test version - all coins are “play” coins and have no value outside of this pilot, you can use them to buy goods during the trail, at the end of it, they will be returned to the system.',
     endpoint: {
-      fullName: 'Welcome to GoodDollar!',
+      fullName: 'Welcome to GoodDollar',
+    },
+  },
+}
+
+export const inviteFriendsMessage = {
+  id: '0',
+  type: 'invite',
+  date: new Date().toString(),
+  status: 'completed',
+  data: {
+    customName: 'Invite friends and earn G$',
+    subtitle: 'Want to earn more G$ ?',
+    receiptData: {
+      from: '0x0000000000000000000000000000000000000000',
+    },
+    reason:
+      'Help expand the network by inviting family, friends, and colleagues to participate and claim their daily income.\nThe more people join, the more effective GoodDollar will be, for everyone.',
+    endpoint: {
+      fullName: 'Invite friends and earn G$',
+    },
+  },
+}
+export const backupMessage = {
+  id: '2',
+  type: 'backup',
+  date: new Date().toString(),
+  status: 'completed',
+  data: {
+    customName: 'Backup your wallet. Now.',
+    subtitle: 'You need to backup your',
+    receiptData: {
+      from: '0x0000000000000000000000000000000000000000',
+    },
+    reason:
+      'Your pass phrase is the only key to your wallet, this is why our wallet is super secure. Only you have access to your wallet and money. But if you won’t backup your pass phrase or if you lose it — you won’t be able to access your wallet and all your money will be lost forever.',
+    endpoint: {
+      fullName: 'Backup your wallet. Now.',
+    },
+  },
+}
+
+export const startSpending = {
+  id: '3',
+  type: 'spending',
+  date: new Date().toString(),
+  status: 'completed',
+  data: {
+    customName: 'Go to GoodMarket',
+    subtitle: 'Start spending your GoodDollars',
+    receiptData: {
+      from: '0x0000000000000000000000000000000000000000',
+    },
+    reason:
+      'Visit GoodMarket, eToro’s exclusive marketplace, where you can buy or sell items in exchange for GoodDollars.',
+    endpoint: {
+      fullName: 'Go to GoodMarket',
     },
   },
 }
@@ -196,6 +280,18 @@ export class UserStorage {
   gun: Gun
 
   /**
+   * a gun node referring tto gun.user().get('properties')
+   * @instance {Gun}
+   */
+  properties: Gun
+
+  /**
+   * a gun node referring tto gun.user().get('properties')
+   * @instance {UserProperties}
+   */
+  userProperties: UserProperties
+
+  /**
    * a gun node refering to gun.user().get('profile')
    * @instance {Gun}
    */
@@ -234,6 +330,8 @@ export class UserStorage {
   subscribersProfileUpdates = []
 
   _lastProfileUpdate: any
+
+  profileSettings: any
 
   /**
    * Magic line for recovery user
@@ -365,6 +463,18 @@ export class UserStorage {
   async init() {
     logger.debug('Initializing GunDB UserStorage')
 
+    this.profileSettings = {
+      fullName: { defaultPrivacy: 'public' },
+      email: { defaultPrivacy: Config.isEToro ? 'public' : 'private' },
+      mobile: { defaultPrivacy: Config.isEToro ? 'public' : 'private' },
+      mnemonic: { defaultPrivacy: 'private' },
+      avatar: { defaultPrivacy: 'public' },
+      walletAddress: { defaultPrivacy: 'public' },
+      username: { defaultPrivacy: 'public' },
+      w3Token: { defaultPrivacy: 'private' },
+      loginToken: { defaultPrivacy: 'private' },
+    }
+
     //sign with different address so its not connected to main user address and there's no 1-1 link
     const username = await this.wallet.sign('GoodDollarUser', 'gundb').then(r => r.slice(0, 20))
     const password = await this.wallet.sign('GoodDollarPass', 'gundb').then(r => r.slice(0, 20))
@@ -399,6 +509,7 @@ export class UserStorage {
       logger.debug('init to events')
 
       await this.initFeed()
+      await this.initProperties()
 
       //save ref to user
       this.gun
@@ -595,7 +706,64 @@ export class UserStorage {
 
     //first time user
     if ((await this.feed) === undefined) {
-      this.enqueueTX(welcomeMessage)
+      const w3Token = await this.getProfileFieldValue('w3Token')
+
+      if (Config.isEToro) {
+        this.enqueueTX(welcomeMessageOnlyEtoro)
+      } else {
+        this.enqueueTX(welcomeMessage)
+      }
+      if (!w3Token) {
+        setTimeout(() => {
+          this.enqueueTX(inviteFriendsMessage)
+        }, 60000)
+      }
+    }
+  }
+
+  /**
+   * Save user properties
+   */
+  async initProperties() {
+    this.properties = this.gunuser.get('properties')
+
+    if ((await this.properties) === undefined) {
+      let putRes = await this.properties.get('properties').put(UserProperties.defaultProperties)
+      logger.debug('set defaultProperties ok:', { defaultProperties: UserProperties.defaultProperties, putRes })
+    }
+    this.userProperties = new UserProperties(this.properties)
+    await this.userProperties.updateLocalData()
+  }
+
+  /**
+   * if necessary, add a backup card
+   *
+   * @returns {Promise<void>}
+   */
+  async addBackupCard() {
+    const userProperties = await this.userProperties.getAll()
+    if (!userProperties.isMadeBackup) {
+      await this.enqueueTX(backupMessage)
+      await this.userProperties.set('isMadeBackup', true)
+    }
+  }
+
+  async onlyForEToro() {
+    if (Config.isEToro) {
+      const userProperties = await this.userProperties.getAll()
+      if (
+        userProperties.firstVisitApp &&
+        Date.now() - userProperties.firstVisitApp >= 68400 &&
+        userProperties.etoroAddCardSpending
+      ) {
+        await this.enqueueTX(startSpending)
+        await this.userProperties.set('etoroAddCardSpending', false)
+      }
+
+      if (!userProperties.firstVisitApp) {
+        await this.userProperties.set('firstVisitApp', Date.now())
+        await this.userProperties.set('etoroAddCardSpending', true)
+      }
     }
   }
 
@@ -675,6 +843,12 @@ export class UserStorage {
     this.subscribersProfileUpdates = []
   }
 
+  async getFieldPrivacy(field) {
+    const currentPrivacy = await this.profile.get(field).get('privacy')
+
+    return currentPrivacy || this.profileSettings[field].defaultPrivacy || 'public'
+  }
+
   /**
    * Save profile with all validations and indexes
    * It saves only known profile fields
@@ -696,26 +870,11 @@ export class UserStorage {
       }
     }
 
-    const profileSettings = {
-      fullName: { defaultPrivacy: 'public' },
-      email: { defaultPrivacy: 'private' },
-      mobile: { defaultPrivacy: 'private' },
-      mnemonic: { defaultPrivacy: 'private' },
-      avatar: { defaultPrivacy: 'public' },
-      walletAddress: { defaultPrivacy: 'public' },
-      username: { defaultPrivacy: 'public' },
-      w3Token: { defaultPrivacy: 'private' },
-      loginToken: { defaultPrivacy: 'private' },
-    }
-    const getPrivacy = async field => {
-      const currentPrivacy = await this.profile.get(field).get('privacy')
-      return currentPrivacy || profileSettings[field].defaultPrivacy || 'public'
-    }
     return Promise.all(
-      keys(profileSettings)
+      keys(this.profileSettings)
         .filter(key => profile[key])
         .map(async field => {
-          return this.setProfileField(field, profile[field], await getPrivacy(field)).catch(e => {
+          return this.setProfileField(field, profile[field], await this.getFieldPrivacy(field)).catch(e => {
             logger.error('setProfile field failed:', { field }, e.message, e)
             return { err: `failed saving field ${field}` }
           })
@@ -1021,6 +1180,39 @@ export class UserStorage {
   }
 
   /**
+   * Save survey
+   * @param {string} hash
+   * @param {object} details
+   * @returns {Promise<void>}
+   */
+  saveSurveyDetails(hash, details: SurveyDetails) {
+    try {
+      const date = moment(new Date()).format('DDMMYY')
+      this.gun
+        .get('survey')
+        .get(date)
+        .get(hash)
+        .put(details)
+      return true
+    } catch (e) {
+      logger.error('saveSurveyDetails :', details, e.message, e)
+      return false
+    }
+  }
+
+  /**
+   * Get all survey
+   * @returns {Promise<void>}
+   */
+  async getSurveyDetailByHashAndDate(hash: string, date: string) {
+    const result = await this.gun
+      .get('survey')
+      .get(date)
+      .get(hash)
+    return result
+  }
+
+  /**
    *
    * @param {string} field - Profile field value (email, mobile or wallet address value)
    * @returns { string } address
@@ -1136,6 +1328,7 @@ export class UserStorage {
           amount: value,
           message: reason || message,
           subtitle,
+          withdrawCode,
         },
       }
     } catch (e) {
@@ -1545,41 +1738,52 @@ export class UserStorage {
    * Calling the server to delete their data
    */
   async deleteAccount(): Promise<boolean> {
-    let deleteResults = await Promise.all([
-      this.wallet
-        .deleteAccount()
-        .then(r => ({ wallet: 'ok' }))
-        .catch(e => ({ wallet: 'failed' })),
-      API.deleteAccount(this.wallet.getAccountForType('zoomId'))
-        .then(r => get(r, 'data.results'))
-        .catch(e => ({
-          server: 'failed',
-        })),
-      this.deleteProfile()
-        .then(r => ({
-          profile: 'ok',
-        }))
-        .catch(r => ({
-          profile: 'failed',
-        })),
-      this.gunuser
-        .get('feed')
-        .putAck(null)
-        .then(r => ({
-          feed: 'ok',
-        }))
-        .catch(r => ({
-          feed: 'failed',
-        })),
-    ])
+    const zoomId = await this.wallet.getAccountForType('zoomId').replace('0x', '')
+    const zoomSignature = await this.wallet.sign(zoomId, 'zoomId')
+    let deleteResults = false
+    let deleteAccountResult
 
-    //Issue with gun delete()
-    // let profileDelete = await this.gunuser
-    //   .delete()
-    //   .then(r => ({ profile: 'ok' }))
-    //   .catch(e => ({
-    //     profile: 'failed'
-    //   }))
+    try {
+      deleteAccountResult = await API.deleteAccount(zoomId, zoomSignature)
+    } catch (e) {
+      logger.error('deleteAccount', { e })
+      return false
+    }
+
+    if (deleteAccountResult.data.ok) {
+      deleteResults = await Promise.all([
+        this.wallet
+          .deleteAccount()
+          .then(r => ({ wallet: 'ok' }))
+          .catch(e => ({ wallet: 'failed' })),
+        this.deleteProfile()
+          .then(r => ({
+            profile: 'ok',
+          }))
+          .catch(r => ({
+            profile: 'failed',
+          })),
+        this.gunuser
+          .get('feed')
+          .putAck(null)
+          .then(r => ({
+            feed: 'ok',
+          }))
+          .catch(r => ({
+            feed: 'failed',
+          })),
+        this.gunuser
+          .get('properties')
+          .putAck(null)
+          .then(r => ({
+            properties: 'ok',
+          }))
+          .catch(r => ({
+            properties: 'failed',
+          })),
+      ])
+    }
+
     logger.debug('deleteAccount', { deleteResults })
     return deleteResults
   }
